add admin

+++ client/src/screens/Admin/ListCourseAdmin.js
@@ -6,6 +6,7 @@ import {
   TouchableOpacity,
   Dimensions,
   Modal,
+  Alert,
 } from 'react-native';
 import React, { useEffect, useState } from 'react';
 import Octicons from 'react-native-vector-icons/Octicons';
@@ -15,6 +16,7 @@ import MaterialIcons from 'react-native-vector-icons/MaterialIcons';
 import AntDesign from 'react-native-vector-icons/AntDesign';
 import Config from 'react-native-config';
 import axios from 'axios';
+import Toast from 'react-native-toast-message';
 
 import EditModal from './Add/EditModal';
 
@@ -24,6 +26,7 @@ export default function ListCourseAdmin({ navigation }) {
   const [showAddCourseModal, setShowAddCourseModal] = useState(false);
   const [showEditNameModal, setShowEditNameModal] = useState(false);
   const [idEdit, setIdEdit] = useState();
+  const [nameEdit, setNameEdit] = useState('');
 
   const [listCourse, setListCourse] = useState([]);
 
@@ -41,17 +44,39 @@ export default function ListCourseAdmin({ navigation }) {
   // ];
 
   const deleteCourse = (id) => {
-    axios.delete(`${Config.API_URL}/courses/${id}`).then((res) => {})
+    axios.delete(`${Config.API_URL}/courses/${id}`).then((res) => {
+      Toast.show({
+        type: 'successToast',
+        text1: 'Xóa khóa học thành công',
+        visibilityTime: 2000,
+      });
+      const filteredCourse = listCourse.filter((course) => course.course_id !== id);
+      setListCourse(filteredCourse);
+    });
   };
 
   const addNewCourse = (newCourseName) => {
     // Xử lý thêm khóa học
     axios.post(`${Config.API_URL}/courses`, { course_name: newCourseName }).then((res) => {
-      setListCourse(...listCourse, {
-        course_id: res.data.course_id,
-        course_name: res.data.course_name,
-        totalChapter: 0,
-      });
+      if (res.data.error) {
+        Toast.show({
+          type: 'errorToast',
+          text1: res.data.error,
+          visibilityTime: 2000,
+        });
+      } else {
+        Toast.show({
+          type: 'successToast',
+          text1: 'Thêm khóa học thành công',
+          visibilityTime: 2000,
+        });
+        const newCourse = {
+          course_id: res.data.course_id,
+          course_name: res.data.course_name,
+          totalChapter: 0,
+        };
+        setListCourse([...listCourse, newCourse]);
+      }
     });
     // Bắn alert
     setShowAddCourseModal(false);
@@ -60,11 +85,23 @@ export default function ListCourseAdmin({ navigation }) {
   const editCourseName = (newName) => {
     // Xử lý thay tên khóa học
     axios.post(`${Config.API_URL}/courses/${idEdit}`, { course_name: newName }).then((res) => {
-      setListCourse(...listCourse, {
-        course_id: res.data.course_id,
-        course_name: res.data.course_name,
-        totalChapter: 0,
-      });
+      if (res.data.error) {
+        Toast.show({
+          type: 'errorToast',
+          text1: res.data.error,
+          visibilityTime: 2000,
+        });
+      } else {
+        Toast.show({
+          type: 'successToast',
+          text1: 'Sửa khóa học thành công',
+          visibilityTime: 2000,
+        });
+        const index = listCourse.findIndex((course) => course.course_id === idEdit);
+        const filteredCourse = [...listCourse];
+        filteredCourse[index].course_name = newName;
+        setListCourse(filteredCourse);
+      }
     });
     // Bắn alert
     // id của khóa học cần thay tên lưu trong state idEdit
@@ -98,6 +135,7 @@ export default function ListCourseAdmin({ navigation }) {
       >
         <EditModal
           editOrAdd="edit course"
+          value={nameEdit}
           setVisible={() => setShowEditNameModal(false)}
           onPressHandle={editCourseName}
         />
@@ -112,7 +150,7 @@ export default function ListCourseAdmin({ navigation }) {
           <TouchableOpacity
             style={styles.course}
             onPress={() =>
-              navigation.navigate('ListChapterAdmin', {
+              navigation.navigate('ChapterAdmin', {
                 courseName: item.course_name,
                 courseId: item.course_id,
               })
@@ -147,6 +185,7 @@ export default function ListCourseAdmin({ navigation }) {
               style={styles.icon}
               onPress={() => {
                 setIdEdit(item.course_id);
+                setNameEdit(item.course_name);
                 setShowEditNameModal(true);
               }}
             >
@@ -155,7 +194,18 @@ export default function ListCourseAdmin({ navigation }) {
             <TouchableOpacity
               style={styles.icon}
               onPress={() => {
-                deleteCourse(item.course_id);
+                Alert.alert('Xóa', 'Bạn có chắc chắn muốn xóa khóa học này', [
+                  {
+                    text: 'Chắc chắn',
+                    style: 'cancel',
+                    onPress: () => {
+                      deleteCourse(item.course_id);
+                    },
+                  },
+                  {
+                    text: 'Hủy',
+                  },
+                ]);
               }}
             >
               <FontAwesome5 name={'trash'} size={16} color={'#ff3636'} />
@@ -215,7 +265,7 @@ const styles = StyleSheet.create({
     paddingBottom: 8,
   },
   chapterCourse: {
-    fontSize: 17,
+    fontSize: 15,
     color: 'white',
     fontWeight: '600',
   },
